factor bot into its own code unit, eliminate global state
